third_party/blink/web_tests/external/wpt/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/legend-expected.txt

This is a testharness.js-based test.
Found 4 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] in-body: unicodeBidi
  assert_equals: expected "isolate" but got "normal"
[FAIL] rendered-legend: unicodeBidi
  assert_equals: expected "isolate" but got "normal"
[FAIL] in-fieldset-second-child: unicodeBidi
  assert_equals: expected "isolate" but got "normal"
[FAIL] in-fieldset-descendant: unicodeBidi
  assert_equals: expected "isolate" but got "normal"
Harness: the test ran to completion.
